feat(deps): migrate to rpc-client-go/v2 and deltabot-cli-go/v2
Drop rpc-client-go v1 and deltabot-cli-go v1 entirely; replace with
their v2 counterparts (v2.49.0), which are versioned in lockstep.

All source imports updated to the /v2 module paths. The domain layer is
also updated to use uint32 for all Delta Chat-origin IDs, matching the
native type in v2 and eliminating the uint64↔uint32 translation layer.
Concretely:

- domain.Messenger interface and IncomingMessage/QuotedMessage fields:
  uint64 → uint32
- adapter/deltachat/messenger.go: drop typed ID wrappers (AccountId etc.),
  rename MsgData→MessageData and MsgType→Viewtype constants,
  wrap SendReaction emoji in []string, replace lastSpecialContactID
  constant with dc.ContactLastSpecial, handle *MessageQuoteWithMessage
  variant for quote extraction
- bot.go: /v2 imports; OnNewMsg callback retyped to uint32
- handler.go: uint64→uint32 cascade; inline convertChatID (uint32 always
  fits int64, no overflow possible); remove errChatIDOverflow, math
  import, and all G115 gosec suppressions
- handler_test.go: uint64→uint32 in mock signatures and fixture IDs;
  remove TestConvertChatID (function no longer exists)

+++ internal/domain/models.go
@@ -65,9 +65,9 @@ const (
 
 // IncomingMessage represents an incoming chat message in domain terms.
 type IncomingMessage struct {
-	ID            uint64
-	ChatID        uint64
-	FromID        uint64
+	ID            uint32
+	ChatID        uint32
+	FromID        uint32
 	Text          string
 	File          string        // local path to attached file, empty if none
 	MediaType     string        // domain media type constant (image/sticker/gif/video), empty if not media
@@ -77,7 +77,7 @@ type IncomingMessage struct {
 
 // QuotedMessage represents the message being replied to.
 type QuotedMessage struct {
-	MessageID uint64
+	MessageID uint32
 }
 
 // FilterResponse represents the response associated with a filter
