+++ internal/domain/ports.go
@@ -41,28 +41,28 @@ type AIClient interface {
 }
 
 // Messenger defines the port for chat messaging operations.
-// All ID parameters use uint64 to match Delta Chat's native types.
+// All ID parameters use uint32 to match Delta Chat's native types.
 type Messenger interface {
 	// FetchMessage retrieves a message by ID.
-	FetchMessage(accountID uint64, msgID uint64) (*IncomingMessage, error)
+	FetchMessage(accountID uint32, msgID uint32) (*IncomingMessage, error)
 	// FetchChatType retrieves the chat type (group/single) for a chat.
-	FetchChatType(accountID uint64, chatID uint64) (ChatType, error)
+	FetchChatType(accountID uint32, chatID uint32) (ChatType, error)
 	// SendTextReply sends a text message as a quote-reply. Returns the sent message ID.
-	SendTextReply(accountID uint64, chatID uint64, replyTo uint64, text string) (uint64, error)
+	SendTextReply(accountID uint32, chatID uint32, replyTo uint32, text string) (uint32, error)
 	// SendMediaReply sends a media file as a quote-reply. mediaType is a domain media type constant.
-	SendMediaReply(accountID uint64, chatID uint64, replyTo uint64, filePath string, mediaType string) (uint64, error)
+	SendMediaReply(accountID uint32, chatID uint32, replyTo uint32, filePath string, mediaType string) (uint32, error)
 	// SendReaction sends a reaction emoji on a message.
-	SendReaction(accountID uint64, msgID uint64, reaction string) error
+	SendReaction(accountID uint32, msgID uint32, reaction string) error
 	// SendTextMessage sends a plain text message (no quote-reply).
-	SendTextMessage(accountID uint64, chatID uint64, text string) error
+	SendTextMessage(accountID uint32, chatID uint32, text string) error
 	// DownloadMessage downloads a message's full media content.
-	DownloadMessage(accountID uint64, msgID uint64) error
+	DownloadMessage(accountID uint32, msgID uint32) error
 	// IsSpecialContact reports whether the given contact ID is a system/device contact
 	// that should be ignored by the bot (e.g. self, device-chat, info bot).
-	IsSpecialContact(fromID uint64) bool
+	IsSpecialContact(fromID uint32) bool
 	// FetchContactDisplayName retrieves the display name for a contact.
 	// Falls back to the contact's name, then email address if display name is empty.
-	FetchContactDisplayName(accountID uint64, contactID uint64) (string, error)
+	FetchContactDisplayName(accountID uint32, contactID uint32) (string, error)
 }
 
 // ConversationRepository defines database operations for conversation threads.
